More x86-64 entry paging refactorings.
