--- source/kernel/arch/x86_64/boot/identity_map.h
@@ -1,8 +1,0 @@
-#ifndef SOS_IDENTITY_MAP_H
-#define SOS_IDENTITY_MAP_H
-
-#include "../../../boot/multiboot.h"
-
-void identity_map_ram(const multiboot_info* const mboot_info);
-
-#endif // SOS_IDENTITY_MAP_H
