+++ source/kernel/arch/x86_64/boot/x86_64_pmm_init.h
@@ -1,0 +1,8 @@
+#ifndef SOS_X86_64_PMM_INIT_H
+#define SOS_X86_64_PMM_INIT_H
+
+#include "../../../boot/multiboot.h"
+
+void pmm_init(const multiboot_info* mboot_info);
+
+#endif // SOS_X86_64_PMM_INIT_H
